Small improvements

+++ src/stages/prepare.py
@@ -15,15 +15,18 @@ class PrepareStage:
             os.makedirs(self.config.paths["build"])
         shutil.copyfile(
             os.path.join(self.config.paths["root"], "mkpm.mk"),
-            os.path.join(self.config.paths["build"], "mkpm.mk")
+            os.path.join(self.config.paths["build"], "mkpm.mk"),
         )
-        if os.path.exists(os.path.join(self.config.paths["root"], ".mkpm")):
+        if os.path.exists(
+            os.path.join(self.config.paths["root"], ".mkpm")
+        ) and not os.path.exists(os.path.join(self.config.paths["build"], ".mkpm")):
             shutil.copytree(
                 os.path.join(self.config.paths["root"], ".mkpm"),
-                os.path.join(self.config.paths["build"], ".mkpm")
+                os.path.join(self.config.paths["build"], ".mkpm"),
             )
         if os.path.exists(self.config.paths["lb"]):
             shutil.rmtree(self.config.paths["lb"])
         shutil.copytree(
             os.path.join(self.config.paths["root"], "lb"), self.config.paths["lb"]
         )
+        os.chdir(self.config.paths["lb"])
